Replace expect/actual crypto with CryptoProvider interface
Collapses the five expect/actual declarations in tweetnacl-multiplatform
(secureRandomBytes, signInternal, isOnCurveInternal, generateKeyInternal,
secretBoxInternal) into a single internal CryptoProvider interface plus
one expect fun platformCryptoProvider() per native target family.

New shape:
- commonMain/.../crypto/CryptoProvider.kt - interface + DefaultCryptoProvider
  (lazy val) + the lone expect fun
- jvmAndroidMain/.../crypto/JvmCryptoProvider.kt - tweetnacl-java +
  BouncyCastle + java.security.SecureRandom impl
- nativeMain/.../crypto/NativeCryptoProviderBase.kt - shared C-interop for
  sign/generateKey/isOnCurve/secretBox; secureRandomBytes is abstract
- appleMain/linuxMain/mingwMain - thin subclasses overriding
  secureRandomBytes for the platform's CSPRNG, plus the actual fun

Public API is unchanged from Kotlin source: TweetNaCl.Signature.sign/
generateKey/isOnCurve, TweetNaCl.SecretBox(key).box/open, and
secureRandomBytes(count) all keep their signatures. The companion bodies
now delegate to DefaultCryptoProvider instead of the removed top-level
expect funs.

JVM-bytecode rename: the SecureRandom function moved from a platform-
specific file (SecureRandom.jvm.kt -> generated class SecureRandom_jvmKt)
to commonMain (SecureRandom.kt -> SecureRandomKt). Source-level callers
are unaffected; Java consumers using the synthetic Kt class name directly
need to update. BCV baseline updated.
