+++ tweetnacl-multiplatform/src/commonMain/kotlin/net/avianlabs/solana/tweetnacl/SecureRandom.kt
@@ -1,12 +1,14 @@
 package net.avianlabs.solana.tweetnacl
 
+import net.avianlabs.solana.tweetnacl.crypto.DefaultCryptoProvider
+
 /**
  * Returns [count] cryptographically secure random bytes.
  *
  * Each platform delegates to the OS-provided CSPRNG:
- * - JVM: `java.security.SecureRandom`
+ * - JVM / Android: `java.security.SecureRandom`
  * - Apple: `SecRandomCopyBytes`
  * - Linux: `/dev/urandom`
  * - Windows: `BCryptGenRandom`
  */
-public expect fun secureRandomBytes(count: Int): ByteArray
+public fun secureRandomBytes(count: Int): ByteArray = DefaultCryptoProvider.secureRandomBytes(count)
