+++ tweetnacl-multiplatform/src/commonMain/kotlin/net/avianlabs/solana/tweetnacl/TweetNaCl.kt
@@ -1,5 +1,6 @@
 package net.avianlabs.solana.tweetnacl
 
+import net.avianlabs.solana.tweetnacl.crypto.DefaultCryptoProvider
 import net.avianlabs.solana.tweetnacl.ed25519.Ed25519Keypair
 
 public interface TweetNaCl {
@@ -46,13 +47,13 @@ public interface TweetNaCl {
       public const val SIGNATURE_BYTES: Int = 64
 
       override fun sign(message: ByteArray, secretKey: ByteArray): ByteArray =
-        signInternal(message, secretKey)
+        DefaultCryptoProvider.sign(message, secretKey)
 
       override fun generateKey(seed: ByteArray): Ed25519Keypair =
-        generateKeyInternal(seed)
+        DefaultCryptoProvider.generateKey(seed)
 
       override fun isOnCurve(publicKey: ByteArray): Boolean =
-        isOnCurveInternal(publicKey)
+        DefaultCryptoProvider.isOnCurve(publicKey)
     }
   }
 
@@ -88,13 +89,7 @@ public interface TweetNaCl {
       public const val KEY_BYTES: Int = 32
 
       public operator fun invoke(secretKey: ByteArray): SecretBox =
-        secretBoxInternal(secretKey)
+        DefaultCryptoProvider.secretBox(secretKey)
     }
   }
 }
-
-internal expect fun signInternal(message: ByteArray, secretKey: ByteArray): ByteArray
-internal expect fun isOnCurveInternal(publicKey: ByteArray): Boolean
-internal expect fun generateKeyInternal(seed: ByteArray): Ed25519Keypair
-
-internal expect fun secretBoxInternal(secretKey: ByteArray): TweetNaCl.SecretBox
