--- tweetnacl-multiplatform/src/jvmAndroidMain/kotlin/net/avianlabs/solana/tweetnacl/SecureRandom.jvm.kt
@@ -1,8 +1,0 @@
-package net.avianlabs.solana.tweetnacl
-
-import java.security.SecureRandom
-
-private val secureRandom = SecureRandom()
-
-public actual fun secureRandomBytes(count: Int): ByteArray =
-  ByteArray(count).also { secureRandom.nextBytes(it) }
